refactor: complicated project structure
The original working branch of this project is a single file and runs just fine. This is really just setup for playing with the idea of a CLI that can fetch JSONs from APIs and present it in a formatted manner to various output paths in an extensible way. Parts of it are still very ugly.